put tx_fee function into libwallet::transaction
